Rename version API files
