+++ src/versions/MinecraftReleaseList.js
